EPMLABSBRN-171 Performed slight refactor of the code structure and added state management
