+++ frontend-angular/package.json
@@ -21,6 +21,10 @@
     "@angular/platform-browser": "8.2.9",
     "@angular/platform-browser-dynamic": "8.2.9",
     "@angular/router": "8.2.9",
+    "@ngrx/effects": "^8.6.0",
+    "@ngrx/entity": "^8.6.0",
+    "@ngrx/store": "^8.6.0",
+    "@ngrx/store-devtools": "^8.6.0",
     "fp-ts": "^2.1.2",
     "ngx-take-until-destroy": "^5.4.0",
     "rxjs": "~6.4.0",
