+++ frontend-angular/src/app/admin/admin-page.component.scss
@@ -6,6 +6,7 @@
   justify-content: center;
   width: 400px;
   height: 200px;
+  border: thin solid black;
 
   & button:not(:last-child) {
     margin-right: 10px;
