+++ frontend-angular/src/app/shared/services/utils.ts
