+++ frontend-angular/src/app/shared/shared.module.ts
@@ -1,10 +1,13 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 
+import { UploadFileComponent } from './components/upload-file/upload-file.component';
+import { UploadService } from './services/upload/upload.service';
 
 @NgModule({
-  declarations: [],
-  exports: [],
+  declarations: [UploadFileComponent],
+  exports: [UploadFileComponent],
+  providers: [UploadService],
   imports: [
     CommonModule
   ]
